requirements.txt
# Component 1 — YouTube Transcript Fetcher
youtube-transcript-api>=0.6.0
yt-dlp>=2024.1.0
faster-whisper>=1.0.0

# Component 2 — Chunker & Embedder
openai>=1.0.0
sentence-transformers>=3.0.0

# Component 3 — Vector Store
chromadb>=0.5.0

# Component 4 — RAG Query Engine
anthropic>=0.40.0

# Component 5 — Frontend UI
streamlit>=1.40.0

# Utilities
python-dotenv>=1.0.0
